SpringBoot-Ticket-Management-System-master/src/main/java/com/practice1/TicketManagement/model/Ticket.java
package com.practice1.TicketManagement.model;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.Id;
import jakarta.persistence.Table;

@Entity
@Table(name="ticket")
public class Ticket {

    @Id
    @Column
    private Integer ticketId;
    @Column
    private String passengerName;
    @Column
    private String source;
    @Column
    private String destination;

    public Ticket() {
    }

    public Ticket(Integer ticketId, String passengerName, String source, String destination) {
        this.ticketId = ticketId;
        this.passengerName = passengerName;
        this.source = source;
        this.destination = destination;
    }

    public Integer getTicketId() {
        return ticketId;
    }

    public void setTicketId(Integer ticketId) {
        this.ticketId = ticketId;
    }

    public String getPassengerName() {
        return passengerName;
    }

    public void setPassengerName(String passengerName) {
        this.passengerName = passengerName;
    }

    public String getSource() {
        return source;
    }

    public void setSource(String source) {
        this.source = source;
    }

    public String getDestination() {
        return destination;
    }

    public void setDestination(String destination) {
        this.destination = destination;
    }
}
